+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/commands.py
@@ -107,6 +107,12 @@ def load_command_table(self, _):
         g.custom_command('set', 'update_connection_strings')
         g.custom_command('delete', 'delete_connection_strings')
 
+    with self.command_group('webapp config storage-account') as g:
+        g.custom_command('list', 'get_azure_storage_accounts', exception_handler=empty_on_404)
+        g.custom_command('add', 'add_azure_storage_account')
+        g.custom_command('update', 'update_azure_storage_account')
+        g.custom_command('delete', 'delete_azure_storage_accounts')
+
     with self.command_group('webapp config hostname') as g:
         g.custom_command('add', 'add_hostname', exception_handler=ex_handler_factory())
         g.custom_command('list', 'list_hostnames')
@@ -179,7 +185,7 @@ def load_command_table(self, _):
         g.custom_command('list-locations', 'list_locations', transform=transform_list_location_output)
 
     with self.command_group('functionapp') as g:
-        g.custom_command('create', 'create_function')
+        g.custom_command('create', 'create_function', exception_handler=ex_handler_factory())
         g.custom_command('list', 'list_function_app', table_transformer=transform_web_list_output)
         g.custom_show_command('show', 'show_webapp', table_transformer=transform_web_output)
         g.custom_command('delete', 'delete_function_app')