+++ src/command_modules/azure-cli-container/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.3.4
++++++
+* Added ability to restart and stop a running container group
+* Add '--network-profile' for passing in a network profile
+
 0.3.3
 +++++
 * Add '--secure-environment-variables' for passing secure environment variables to a container